fix(review): decompose protocols.py, drop config<->profiles circular re-export, clarify docs
- Move EventLog, ArtifactStore, EpisodicStore, FactStore from protocols.py
  into a new src/contextweaver/store/protocols.py (protocols.py is now
  192 lines, store/protocols.py 230 lines — both under the <=300 guideline).
  The four store protocols remain importable from contextweaver.protocols and
  contextweaver via aliased re-exports for backward compatibility.

- Remove the bottom-of-file `from contextweaver.profiles import ...` re-export
  in config.py to resolve the circular-import smell. RoutingConfig and
  ProfileConfig are now imported from contextweaver.profiles or contextweaver.
  Internal callers migrated: routing/router.py, tests/test_router.py,
  tests/test_config.py.

- StoreBundle.to_dict()/from_dict() docstrings spell out the silent-None
  round-trip behaviour for custom backends that lack a to_dict() method.

- ProfileConfig.to_dict() docstring expanded to note policy is included now
  that ContextPolicy is serialisable, and that the round-trip is lossless.

AGENTS.md module map and CHANGELOG.md [Unreleased] updated.
make ci: fmt + lint + test (621 passed) + example + demo all pass; mypy
retains the one pre-existing fastmcp.py:252 error noted in the PR description.

Addresses dgenio self-review must-fix and three should-fix items, plus the
five Copilot inline comments (four already resolved on the prior commit).
