+++ CHANGELOG.md
@@ -14,18 +14,30 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `to_dict()` / `from_dict()` on `ContextPolicy`, `ContextBudget`, and
   `ScoringConfig` — completes the repo-standard serialisation methods on all
   config dataclasses (#184)
-- `EpisodicStore` and `FactStore` protocols in `protocols.py` — formal
-  `@runtime_checkable` protocol interfaces matching the `InMemory*` method
-  signatures; `StoreBundle` type hints widened to protocol types (#40)
+- `EpisodicStore` and `FactStore` protocols — formal `@runtime_checkable`
+  protocol interfaces matching the `InMemory*` method signatures; `StoreBundle`
+  type hints widened to protocol types (#40)
+- `store/protocols.py` module — store-layer protocols (`EventLog`,
+  `ArtifactStore`, `EpisodicStore`, `FactStore`) extracted from `protocols.py`
+  to stay within the ≤300-line guideline; still importable from
+  `contextweaver.protocols` and `contextweaver` for backward compatibility
 - `profiles.py` module — `RoutingConfig` and `ProfileConfig` extracted from
-  `config.py` to stay within the ≤300-line guideline; re-exported from
-  `config.py` for backward compatibility (#179)
+  `config.py` to stay within the ≤300-line guideline; importable from
+  `contextweaver.profiles` and `contextweaver` (#179)
 
 ### Changed
 - `ProfileConfig.to_dict()` / `from_dict()` now include `policy` (previously
-  excluded because `ContextPolicy` lacked serialisation) (#184)
+  excluded because `ContextPolicy` lacked serialisation); docstring expanded
+  to make the round-trip contract explicit (#184)
 - `ContextManager.episodic_store` / `fact_store` properties now return protocol
   types (`EpisodicStore` / `FactStore`) instead of concrete `InMemory*` types (#40)
+- `StoreBundle.to_dict()` / `from_dict()` docstrings now spell out the
+  silent-`None` round-trip behaviour for custom backends that lack a
+  `to_dict()` method
+- `RoutingConfig` / `ProfileConfig` are no longer re-exported from
+  `contextweaver.config`; import them from `contextweaver.profiles` or
+  `contextweaver`. Dropping the re-export resolves a circular-import smell
+  between `config.py` and `profiles.py`
 
 ### Fixed
 - Replaced 3 bare `ValueError` raises in `context/sensitivity.py` with
